Safari-609.1.11.1/LayoutTests/js/for-in-avoid-duplicates-expected.txt

This tests that for/in statements don't report properties that are in both an object and its prototype more than once.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS i is 2
PASS successfullyParsed is true

TEST COMPLETE
